+++ persistent/Database/Persist/Class/PersistQuery.hs
@@ -2,21 +2,24 @@
 module Database.Persist.Class.PersistQuery
     ( PersistQueryRead (..)
     , PersistQueryWrite (..)
+    , PersistQueryStream (..)
     , selectSource
     , selectKeys
     , selectList
     , selectKeysList
+    , selectStream
     ) where
 
 import Control.Monad.IO.Class (MonadIO, liftIO)
-import Control.Monad.Reader   (ReaderT, MonadReader)
+import Control.Monad.Reader (MonadReader(ask), ReaderT, lift)
 import Control.Monad.Trans.Resource (MonadResource, release)
 import Data.Acquire (Acquire, allocateAcquire, with)
-import Data.Conduit (ConduitM, (.|), await, runConduit)
+import Data.Conduit (ConduitM, await, runConduit, (.|))
+import Data.Conduit.Lift (runReaderC)
 import qualified Data.Conduit.List as CL
 
-import Database.Persist.Class.PersistStore
 import Database.Persist.Class.PersistEntity
+import Database.Persist.Class.PersistStore
 
 -- | Backends supporting conditional read operations.
 class (PersistCore backend, PersistStoreRead backend) => PersistQueryRead backend where
@@ -53,6 +56,31 @@ class (PersistCore backend, PersistStoreRead backend) => PersistQueryRead backen
     -- @since 2.11
     exists :: (MonadIO m, PersistRecordBackend record backend)
            => [Filter record] -> ReaderT backend m Bool
+
+-- | Backends supporting conditional read operations, which may also support
+-- streaming the results in a memory-constant way.
+--
+-- @since 2.13.2.0
+class (PersistQueryRead backend) => PersistQueryStream backend where
+    -- | Get all records matching the given criterion in the specified order.
+    --
+    -- A version of 'selectSourceRes' which specifically streams the results,
+    -- for SQL backends that support it. Streaming may be slower for small
+    -- query sets, but avoids allocating all the results in memory at once.
+    --
+    -- By default, this behaves identically to 'selectSourceRes'.
+    --
+    -- @since 2.13.2.0
+    selectSourceStream
+        :: (PersistRecordBackend record backend, MonadResource m)
+        => [Filter record]
+        -> [SelectOpt record]
+        -> ConduitM () (Entity record) (ReaderT backend m) ()
+    selectSourceStream filts opts = do
+      srcRes <- lift $ selectSourceRes filts opts
+      (releaseKey, src) <- allocateAcquire srcRes
+      src
+      release releaseKey
 
 -- | Backends supporting conditional write operations
 class (PersistQueryRead backend, PersistStoreWrite backend) => PersistQueryWrite backend where
@@ -105,3 +133,14 @@ selectKeysList :: forall record backend m. (MonadIO m, PersistQueryRead backend,
 selectKeysList filts opts = do
     srcRes <- selectKeysRes filts opts
     liftIO $ with srcRes (\src -> runConduit $ src .| CL.consume)
+
+-- | Get all records matching the given criterion in the specified order. Uses
+-- streaming where possible, based on the 'PersistQueryStream' instance for the
+-- backend.
+--
+-- @since 2.13.2.0
+selectStream :: forall record backend m. (PersistQueryStream backend, MonadResource m, PersistRecordBackend record backend, MonadReader backend m)
+             => [Filter record]
+             -> [SelectOpt record]
+             -> ConduitM () (Entity record) m ()
+selectStream filts opts = lift ask >>= (`runReaderC` selectSourceStream filts opts)