+++ persistent/Database/Persist/SqlBackend.hs
@@ -22,19 +22,22 @@ module Database.Persist.SqlBackend
     , setConnInsertManySql
     , setConnUpsertSql
     , setConnPutManySql
+    , setConnPrepareCursor
     ) where
 
 import Control.Monad.Reader
+import Data.List.NonEmpty (NonEmpty)
 import Data.Text (Text)
 import Database.Persist.Class.PersistStore (BackendCompatible(..))
+import Database.Persist.Names
 import Database.Persist.SqlBackend.Internal
 import qualified Database.Persist.SqlBackend.Internal as SqlBackend
        (SqlBackend(..))
-import Database.Persist.SqlBackend.Internal.MkSqlBackend as Mk (MkSqlBackendArgs(..))
-import Database.Persist.Types.Base
-import Database.Persist.Names
 import Database.Persist.SqlBackend.Internal.InsertSqlResult
-import Data.List.NonEmpty (NonEmpty)
+import Database.Persist.SqlBackend.Internal.MkSqlBackend as Mk
+       (MkSqlBackendArgs(..))
+import Database.Persist.SqlBackend.Internal.Statement
+import Database.Persist.Types.Base
 
 -- $utilities
 --
@@ -188,3 +191,9 @@ setConnPutManySql
     -> SqlBackend
 setConnPutManySql  mkQuery sb =
     sb { connPutManySql = Just mkQuery }
+
+setConnPrepareCursor
+    :: (Text -> IO Statement)
+    -> SqlBackend
+    -> SqlBackend
+setConnPrepareCursor f sb = sb { connPrepareCursor = Just f }